// front/src/backend_api/backend/models/FfttPlayer.ts
/* tslint:disable */
/* eslint-disable */
/**
 * USKB Tournament
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 3.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Gender } from './Gender';
import {
    GenderFromJSON,
    GenderFromJSONTyped,
    GenderToJSON,
    GenderToJSONTyped,
} from './Gender';

/**
 * 
 * @export
 * @interface FfttPlayer
 */
export interface FfttPlayer {
    /**
     * 
     * @type {string}
     * @memberof FfttPlayer
     */
    licenceNo: string;
    /**
     * 
     * @type {string}
     * @memberof FfttPlayer
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof FfttPlayer
     */
    lastName: string;
    /**
     * 
     * @type {string}
     * @memberof FfttPlayer
     */
    club: string;
    /**
     * 
     * @type {Gender}
     * @memberof FfttPlayer
     */
    gender: Gender;
    /**
     * 
     * @type {number}
     * @memberof FfttPlayer
     */
    nbPoints: number;
}



/**
 * Check if a given object implements the FfttPlayer interface.
 */
export function instanceOfFfttPlayer(value: object): value is FfttPlayer {
    if (!('licenceNo' in value) || value['licenceNo'] === undefined) return false;
    if (!('firstName' in value) || value['firstName'] === undefined) return false;
    if (!('lastName' in value) || value['lastName'] === undefined) return false;
    if (!('club' in value) || value['club'] === undefined) return false;
    if (!('gender' in value) || value['gender'] === undefined) return false;
    if (!('nbPoints' in value) || value['nbPoints'] === undefined) return false;
    return true;
}

export function FfttPlayerFromJSON(json: any): FfttPlayer {
    return FfttPlayerFromJSONTyped(json, false);
}

export function FfttPlayerFromJSONTyped(json: any, ignoreDiscriminator: boolean): FfttPlayer {
    if (json == null) {
        return json;
    }
    return {
        
        'licenceNo': json['licenceNo'],
        'firstName': json['firstName'],
        'lastName': json['lastName'],
        'club': json['club'],
        'gender': GenderFromJSON(json['gender']),
        'nbPoints': json['nbPoints'],
    };
}

export function FfttPlayerToJSON(json: any): FfttPlayer {
    return FfttPlayerToJSONTyped(json, false);
}

export function FfttPlayerToJSONTyped(value?: FfttPlayer | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'licenceNo': value['licenceNo'],
        'firstName': value['firstName'],
        'lastName': value['lastName'],
        'club': value['club'],
        'gender': GenderToJSON(value['gender']),
        'nbPoints': value['nbPoints'],
    };
}